requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=5.1.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
urllib3>=2.2.0